Update script.user.js
PayPal updated their .svg and the background color and broke the script. This update fixes that. I couldn't get the custom icon to be less blurry and I couldn't get it back on the left side of text because I am not very good and this is all spaghetti code.